completed

+++ app/Http/Controllers/API/UserController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\User;
 use Hash;
+use Gate;
 class UserController extends Controller
 {
     /**
@@ -16,6 +17,7 @@ class UserController extends Controller
     public function __construct()
     {
         $this->middleware('auth:api');
+
     }
     /**
      * Display a listing of the resource.
@@ -24,7 +26,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // $this->authorize('isAdmin');
+
+        if (Gate::allows('isAdmin') || Gate::allows ('isAuthor')) {
+            return User::latest()->paginate(10);
+        }
+
     }
 
     /**
@@ -139,8 +146,24 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
+
+
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'User Deleted'];
+    }
+
+    public function search(){
+        if ($search = \Request::get('q')){
+            $users = User::where(function($query) use ($search){
+                $query->where('name', 'LIKE', "%$search%")
+                      ->orWhere('email', 'LIKE', "%$search%")
+                      ->orWhere('type', 'LIKE', "%$search%");
+            })->paginate(20);
+        }else {
+          $users = User::latest()->paginate(5);
+        }
+        return $users;
     }
 }
